Time: 3 ms (81.37%), Space: 56.4 MB (45.12%) - LeetHub

+++ 0643-maximum-average-subarray-i/0643-maximum-average-subarray-i.java
@@ -10,7 +10,7 @@ class Solution {
         maxavg=avg;
         for(i=k;i<nums.length;i++)
         {
-            sum=(double)(sum+nums[i]-nums[c]);
+            sum=sum+nums[i]-nums[c];
             c++;
             avg=sum/k;
             if(maxavg<avg)
